lib: fix Win compile warning

+++ lib/str_util.cpp
@@ -772,7 +772,7 @@ vector<string> split(string s, char delim) {
 // - can't have ..
 //
 bool is_valid_filename(const char* name) {
-    int n = strlen(name);
+    size_t n = strlen(name);
     for (int i=0; i<n; i++) {
         if (iscntrl(name[i])) {
             return false;
